Split push_userdata to improve both code size and performance
Since monomorphization will generate a function for each different L, this lets us have it generate small wrappers that call the same bigger function if there are multiple L used for a single type. It also now "caches" the Lua context, avoiding the reference to reference to reference mess that sometimes happened before.