Avoid extra work in PushGuard's AsLua / AsMutLua
As we already have the raw lua context around we might as well use it. There's nothing that guarantees that we'll call as_lua / as_mut_lua for each access (and indeed we don't in other places), so it's unnecessary for each as_lua / as_mut_lua call to go through the trait.